Cleaning up MSVC warning in unit tests for nrc

+++ modules/core/unit/container/memory/array_as_buffer.cpp
@@ -75,7 +75,7 @@ NT2_TEST_CASE( std_array_1D_as_buffer )
   // array type supports R/W access through Position
   //////////////////////////////////////////////////////////////////////////////
   for(pos[0]=0;pos[0]<5;++pos[0]) 
-    dereference<1UL>(tab,pos) = 10*(1+pos[0]);
+    dereference<1UL>(tab,pos) = double(10*(1+pos[0]));
 
   for(pos[0]=0;pos[0]<5;++pos[0])
     NT2_TEST_EQUAL(dereference<1UL>(tab,pos), 10*(1+pos[0]) );
@@ -106,7 +106,7 @@ NT2_TEST_CASE( std_array_2D_as_buffer )
   //////////////////////////////////////////////////////////////////////////////
   for(pos[1]=0;pos[1]<2;++pos[1])
     for(pos[0]=0;pos[0]<5;++pos[0])
-      dereference<2UL>(tab,pos) = 10*(1+pos[1]) + (1+pos[0]);
+      dereference<2UL>(tab,pos) = double(10*(1+pos[1]) + (1+pos[0]));
 
   for(pos[1]=0;pos[1]<2;++pos[1])
     for(pos[0]=0;pos[0]<5;++pos[0])
@@ -140,7 +140,7 @@ NT2_TEST_CASE( std_array_3D_as_buffer )
   for(pos[2]=0;pos[2]<3;++pos[2])
     for(pos[1]=0;pos[1]<2;++pos[1])
       for(pos[0]=0;pos[0]<2;++pos[0])
-      dereference<3UL>(tab,pos) = 100*(1+pos[2]) + 10*(1+pos[1]) + (1+pos[0]);
+      dereference<3UL>(tab,pos) = double(100*(1+pos[2]) + 10*(1+pos[1]) + (1+pos[0]));
 
   for(pos[2]=0;pos[2]<3;++pos[2])
     for(pos[1]=0;pos[1]<2;++pos[1])
